[BUGFIX] Ensure `phpunit/phpunit v10` compatibility
`typo3/testing-framework` allows the use of `phpunit/phpunit`
in version 9 and 10. If an extension or project does not have
a specific version pinned, the highest gets installed. This
happend to our testing.

To be `phpunit v10` combatible we only have to change the
dataprovider to static methods, which this change does.
`phpunit` version is required to allow phpunit v9 and v10
in parallel, as v10 has a minimum requirement of php8.1.

v9 is used for TYPO3 core v11 testing, and v10 for TYPO3
core v12 to avoid mixing based on php version.

`phpunit v10` enforces a migration of the configuration
file which is not backwards compatible. To mitigate this,
we split the unit and functionl configuration files per
core version and migrate the v12 once to the new format.
This is an additional point why not to mix based on php
version.

Used command(s):

```shell
composer require --no-update --dev \
  phpunit/phpunit:"^9.6.4 || ^10"
```